add observer to count shown movies
