+++ BAZProject/Modules/MovieDetail/MovieDetailViewController.swift
@@ -17,7 +17,10 @@ class MovieDetailViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         self.presenter?.notifyViewLoaded()
-        // Do any additional setup after loading the view.
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        self.presenter?.notifyViewAppeared()
     }
 }
